refactor: relocate register_submit.dart from register screen to actions for cleaner architecture
